+++ lib/src/extensions/flutter_extensions/flutter_extensions.dart
@@ -1,5 +1,7 @@
 export 'align.dart';
+export 'carousel_controller.dart';
 export 'colors/colors.dart';
+export 'directionality.dart';
 export 'focus_scope_extensions.dart';
 export 'future.dart';
 export 'list_widget.dart';